# frontend/src/api/openapi.merged.yaml
openapi: 3.0.0
info:
  title: ApeRAG API
  description: ApeRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /apikeys:
    get:
      summary: Get API keys
      description: Get API keys
      responses:
        '200':
          description: API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apiKeyList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    post:
      summary: Create API key
      description: Create API key
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/apiKeyCreate'
      responses:
        '200':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apiKey'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /apikeys/{apikey_id}:
    put:
      summary: Update API key
      description: Update API key
      security:
        - BearerAuth: []
      parameters:
        - name: apikey_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/apiKeyUpdate'
      responses:
        '200':
          description: API key updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apiKey'
        '401':
          description: Unauthorized
    delete:
      summary: Delete API key
      description: Delete API key
      security:
        - BearerAuth: []
      parameters:
        - name: apikey_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: API key not found
...
  /audit-logs:
    get:
      tags:
        - audit
      summary: List audit logs
      description: List audit logs with filtering options
      operationId: list_audit_logs
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by user ID
        - name: username
          in: query
          required: false
          schema:
            type: string
          description: Filter by username
        - name: resource_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - collection
              - document
              - bot
              - chat
              - message
              - api_key
              - llm_provider
              - llm_provider_model
              - model_service_provider
              - user
              - config
          description: Filter by resource type
        - name: resource_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by resource ID
        - name: api_name
          in: query
          required: false
          schema:
            type: string
          description: Filter by API name
        - name: http_method
          in: query
          required: false
          schema:
            type: string
            enum:
              - POST
              - PUT
              - DELETE
          description: Filter by HTTP method
        - name: status_code
          in: query
          required: false
          schema:
            type: integer
          description: Filter by status code
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by start date
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by end date
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 5000
            default: 1000
          description: Maximum number of records
      responses:
        '200':
          description: Audit logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auditLogList'
        '403':
          description: Admin access required
  /audit-logs/{audit_id}:
    get:
      tags:
        - audit
      summary: Get audit log detail
      description: Get a specific audit log by ID
      operationId: get_audit_log
      parameters:
        - name: audit_id
          in: path
          required: true
          schema:
            type: string
          description: Audit log ID
      responses:
        '200':
          description: Audit log retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auditLog'
        '403':
          description: Admin access required
        '404':
          description: Audit log not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    pageResult:
      description: PageResult info
      properties:
        page_number:
          type: integer
          description: The page number
        page_size:
          type: integer
          description: The page size
        count:
          type: integer
          description: The total count of items
      type: object
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    apiKey:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
    apiKeyList:
      type: object
      description: A list of API keys
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/apiKey'
        pageResult:
          $ref: '#/components/schemas/pageResult'
    apiKeyCreate:
      type: object
      properties:
        description:
          type: string
    apiKeyUpdate:
      type: object
      properties:
        description:
          type: string
...
    auditLog:
      type: object
      description: Audit log entry
      properties:
        id:
          type: string
          description: Audit log ID
        user_id:
          type: string
          nullable: true
          description: User ID who performed the action
        username:
          type: string
          nullable: true
          description: Username for display
        resource_type:
          type: string
          enum:
            - collection
            - document
            - bot
            - chat
            - message
            - api_key
            - llm_provider
            - llm_provider_model
            - model_service_provider
            - user
            - config
          nullable: true
          description: Type of resource
        resource_id:
          type: string
          nullable: true
          description: ID of the resource (extracted at query time)
        api_name:
          type: string
          description: API operation name
        http_method:
          type: string
          description: HTTP method (POST, PUT, DELETE)
        path:
          type: string
          description: API path
        status_code:
          type: integer
          nullable: true
          description: HTTP status code
        start_time:
          type: integer
          format: int64
          description: Request start time (milliseconds since epoch)
        end_time:
          type: integer
          format: int64
          nullable: true
          description: Request end time (milliseconds since epoch)
        duration_ms:
          type: integer
          nullable: true
          description: Request duration in milliseconds (calculated)
        request_data:
          type: string
          nullable: true
          description: Request data (JSON string)
        response_data:
          type: string
          nullable: true
          description: Response data (JSON string)
        error_message:
          type: string
          nullable: true
          description: Error message if failed
        ip_address:
          type: string
          nullable: true
          description: Client IP address
        user_agent:
          type: string
          nullable: true
          description: User agent string
        request_id:
          type: string
          description: Request ID for tracking
        created:
          type: string
          format: date-time
          description: Created timestamp
    auditLogList:
      type: object
      description: List of audit logs
      properties:
        items:
          type: array
          description: Audit log entries
          items:
            $ref: '#/components/schemas/auditLog'
...